Refactor: Consolidate DistanceMetric and clean core namespace
- Consolidate `DistanceMetric` enum in `src/core/vector/metric.rs`.
- Add `Haversine`, `Hamming`, `Tanimoto` variants to `core::vector::DistanceMetric`.
- Move `to_u8` and `from_u8` serialization methods to `core`.
- Update `src/index/vector/mod.rs` to use `core::vector::DistanceMetric`.
- Remove `vector` re-exports from `src/core/mod.rs` to prevent namespace pollution.
- Document changes in `.jules/atlas.md`.
