+++ src/index/vector/mod.rs
@@ -198,96 +198,7 @@ impl PartialEq for CustomMetric {
 /// Type alias for temporal search results: Vec<(timestamp, Vec<(node_id, similarity)>)>
 pub type TemporalSearchResults = Vec<(Timestamp, Vec<(NodeId, f32)>)>;
 
-/// Distance metric used for similarity computation.
-///
-/// Different metrics are suitable for different use cases:
-/// - **Cosine**: Measures angle between vectors, ignores magnitude (semantic similarity)
-/// - **Euclidean**: Measures straight-line distance (spatial data, clustering)
-/// - **DotProduct**: Inner product, preserves magnitude (MaxSim, ColBERT)
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DistanceMetric {
-    /// Cosine similarity: measures angle between vectors, range [-1, 1]
-    Cosine,
-    /// Euclidean distance (L2): measures straight-line distance, range [0, ∞)
-    Euclidean,
-    /// Dot product: inner product of vectors, range (-∞, ∞)
-    DotProduct,
-    /// Haversine: great circle distance for geographic coordinates
-    Haversine,
-    /// Hamming: bit-level distance for binary vectors
-    Hamming,
-    /// Tanimoto: bit-level Jaccard similarity for chemical fingerprints
-    Tanimoto,
-}
-impl DistanceMetric {
-    /// Encode distance metric as a byte for serialization.
-    ///
-    /// Encoding:
-    /// - 0 = Cosine
-    /// - 1 = Euclidean
-    /// - 2 = DotProduct
-    /// - 3 = Haversine
-    /// - 4 = Hamming
-    /// - 5 = Tanimoto
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use aletheiadb::index::vector::DistanceMetric;
-    ///
-    /// assert_eq!(DistanceMetric::Cosine.to_u8(), 0);
-    /// assert_eq!(DistanceMetric::Euclidean.to_u8(), 1);
-    /// assert_eq!(DistanceMetric::DotProduct.to_u8(), 2);
-    /// assert_eq!(DistanceMetric::Haversine.to_u8(), 3);
-    /// assert_eq!(DistanceMetric::Hamming.to_u8(), 4);
-    /// assert_eq!(DistanceMetric::Tanimoto.to_u8(), 5);
-    /// ```
-    pub fn to_u8(self) -> u8 {
-        match self {
-            DistanceMetric::Cosine => 0,
-            DistanceMetric::Euclidean => 1,
-            DistanceMetric::DotProduct => 2,
-            DistanceMetric::Haversine => 3,
-            DistanceMetric::Hamming => 4,
-            DistanceMetric::Tanimoto => 5,
-        }
-    }
-
-    /// Decode distance metric from a byte.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the byte value is not a valid metric encoding (>= 6).
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use aletheiadb::index::vector::DistanceMetric;
-    ///
-    /// assert_eq!(DistanceMetric::from_u8(0).unwrap(), DistanceMetric::Cosine);
-    /// assert_eq!(DistanceMetric::from_u8(1).unwrap(), DistanceMetric::Euclidean);
-    /// assert_eq!(DistanceMetric::from_u8(2).unwrap(), DistanceMetric::DotProduct);
-    /// assert_eq!(DistanceMetric::from_u8(3).unwrap(), DistanceMetric::Haversine);
-    /// assert_eq!(DistanceMetric::from_u8(4).unwrap(), DistanceMetric::Hamming);
-    /// assert_eq!(DistanceMetric::from_u8(5).unwrap(), DistanceMetric::Tanimoto);
-    /// assert!(DistanceMetric::from_u8(6).is_err());
-    /// ```
-    pub fn from_u8(value: u8) -> Result<Self> {
-        match value {
-            0 => Ok(DistanceMetric::Cosine),
-            1 => Ok(DistanceMetric::Euclidean),
-            2 => Ok(DistanceMetric::DotProduct),
-            3 => Ok(DistanceMetric::Haversine),
-            4 => Ok(DistanceMetric::Hamming),
-            5 => Ok(DistanceMetric::Tanimoto),
-            _ => Err(crate::core::error::StorageError::CorruptedData(format!(
-                "Invalid distance metric encoding: {}",
-                value
-            ))
-            .into()),
-        }
-    }
-}
+pub use crate::core::vector::DistanceMetric;
 
 /// Trait for vector indexes supporting approximate k-nearest neighbor search.
 ///
